[#26] Switch back to Session module
The code seemed more complex with Authenticator than it does without. I
found myself checking the source code of Authenticator again and again
to figure out exactly what it was doing.

In particular, I would prefer to use Phoenix's built-in
`action_fallback` feature than introduce a new `fallback` location
inside an "authenticator" module.

The amount of code needed to manage the plug session is really minimal,
so it makes sense to just include it in the generated output.

If nothing else, this will make it easier to reason about than it was
with Authenticator.